Fix entity mapping errors.

+++ src/main/java/org/fs/rallyroundbackend/entity/users/DepartmentEntity.java
@@ -1,5 +1,6 @@
 package org.fs.rallyroundbackend.entity.users;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -20,7 +21,7 @@ import lombok.Setter;
 public class DepartmentEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private short id;
+    private int id;
 
     private String name;
 }
